Alterado o resultado da mediana e da media

+++ ExemploIntefaces/src/exemplo/intefaces/CalculaMedia.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 public class CalculaMedia implements ICalcula{        
     @Override
-    public double calcula(List<Integer> lista){        
-        int total=0;
+    public double calcula(List<Integer> lista){      
+        double total=0;
         for(Integer item : lista){
             total+= item;
         }
